+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.29.2 (??)
 - Feature: ignore template `blockquote`
 - Fix: merge suggestions with same text and empty comments
+- Fix: simplify the forbidden chars in an XML tag to be capture as an immutable
 
 ## 2.29.1 (12 Aug 2024)
 - Feature: take into account composed misspellings in the uppercase finder